saml: use /token instead of /saml/token
