+++ wazo_auth/plugins/http/saml/http.py
@@ -3,18 +3,14 @@
 
 
 import logging
-from typing import Optional
 
-import marshmallow
 from flask import Response, redirect, request
 from saml2.httputil import ServiceError
 from saml2.response import VerificationError
 from saml2.s_utils import UnknownPrincipal, UnsupportedBinding
 from saml2.sigver import SignatureError
 
-from wazo_auth import exceptions, http
-
-from .schemas import SAMLSessionIdSchema
+from wazo_auth import http
 
 logger = logging.getLogger(__name__)
 
@@ -27,14 +23,12 @@ class SAMLACS(http.ErrorCatchingResource):
         auth_service,
         saml_service,
         config,
-        backend_proxy,
     ):
         self._token_service = token_service
         self._user_service = user_service
         self._auth_service = auth_service
         self._saml_service = saml_service
         self._config = config
-        self._backend_proxy = backend_proxy
 
     def post(self):
         try:
@@ -75,14 +69,12 @@ class SAMLSSO(http.ErrorCatchingResource):
         auth_service,
         saml_service,
         config,
-        wazo_user_backend,
     ):
         self._token_service = token_service
         self._user_service = user_service
         self._auth_service = auth_service
         self._saml_service = saml_service
         self._config = config
-        self._backend = wazo_user_backend
 
     def post(self):
         try:
@@ -97,58 +89,4 @@ class SAMLSSO(http.ErrorCatchingResource):
             return Response(
                 status=500,
                 response='SAML configuration missing or SAML client init failed',
-            )
-
-
-class SAMLTOKEN(http.ErrorCatchingResource):
-    def __init__(
-        self,
-        token_service,
-        user_service,
-        auth_service,
-        saml_service,
-        config,
-        wazo_user_backend,
-    ):
-        self._token_service = token_service
-        self._user_service = user_service
-        self._auth_service = auth_service
-        self._saml_service = saml_service
-        self._config = config
-        self._backend = wazo_user_backend
-
-    def post(self):
-        try:
-            args = SAMLSessionIdSchema().load(request.get_json())
-        except marshmallow.ValidationError as e:
-            raise exceptions.UserParamException.from_errors(e.messages)
-
-        if args.get('saml_session_id'):
-            login: Optional[str] = self._saml_service.getUserLogin(
-                args.get('saml_session_id')
-            )
-            backend_name = 'wazo_user'
-            args = {
-                'user_agent': request.headers.get('User-Agent', ''),
-                'mobile': False,
-                'remote_addr': request.remote_addr,
-                'backend': backend_name,
-            }
-            token = self._token_service.new_token(
-                self._backend[backend_name].obj, login[0], args
-            )
-            redacted_token_id = 'XXXXXXXX-XXXX-XXXX-XXXX-XXXX' + token.token[-8:]
-            logger.info(
-                'Successful login: %s got token %s from %s using agent "%s"',
-                login,
-                redacted_token_id,
-                args['remote_addr'],
-                args['user_agent'],
-            )
-
-            return {'data': token.to_dict()}, 200
-        else:
-            return Response(
-                status=400,
-                response='Session id missing',
             )
