+++ wazo_auth/plugins/http/tokens/api.yml
@@ -73,6 +73,12 @@ paths:
                 The `refresh_token` can be used to get a new access token without using the username/password.
                 This is useful for client application that should not store the username and password once the
                 user has logged in a first time.
+            saml_session_id:
+              type: string
+              description: |
+                The `saml_session_id` can be used to get a new access token or refresh token instead of using
+                the username/password. It can be created using the `/saml/sso` and `/saml/acs` resources. A token
+                created this way has the same properties as another token created using this API.
       responses:
         '200':
           description: "The created token's data"
